Partially handle screen rotation using remapping
Use row and column remapping to change either row or column addressing
for each orientation.  This removes the need to reverse the drawing
direction of either x or y in the drawing routines.  However in the 90
and 270 case x and y must still be swapped.

Also, rename the SSD1306 remapping command names so they are more
descriptive.